backend/requirements-core.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
Flask-JWT-Extended>=4.6.0
Flask-Limiter>=3.5.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.0
limits>=3.11.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
requests>=2.31.0
pydantic>=2.11.0
gunicorn>=21.2.0
redis>=5.0.0
rq>=1.15.1
alembic==1.13.2
